requirements.txt
pyserial=3.5
pure-python-adb=0.3.0.dev0
art=5.1